fix: gradle test execution
