+++ services/tsd-api-mock/build.gradle.kts
@@ -10,6 +10,7 @@ version = "1.0-SNAPSHOT"
 dependencies {
     runtimeOnly("com.h2database:h2")
     implementation("org.apache.commons:commons-lang3:3.14.0")
+    testRuntimeOnly("org.junit.platform:junit-platform-launcher")
 }
 
 tasks.test {
